Refactor into useHomeWorld
